Crate f4

Board Support Crate for STMicroelectronics NUCLEO-F411RE and NUCLEO-F401RE

Usage

Follow cortex-m-quickstart instructions but remove the memory.x linker script and the build.rs build script file as part of the configuration of the quickstart crate.

Re-exports

pub extern crate stm32f40x;

pub use adc::Adc;

pub use adc::AdcChannel;

pub use capture::Capture;

pub use i2c::I2c;

pub use lsm9ds1::ImuSettings;

pub use lsm9ds1::Lsm9ds1;

pub use madgwick_ahrs::MadgwickAhrs;

pub use math_utils::Quaternion;

pub use math_utils::Vector3;

pub use pwm::Pwm;

pub use serial::Serial;

pub use serial::U8Writer;

pub use spi::Spi;

pub use timer::Channel;

pub use timer::Timer;

Modules

adc
Analog to Digital Converter (ADC)

button
User button PC13

capture
Input capture interface

clock
System clocking

dma
Direct Memory Access (DMA)

dwt
Methods for trying a closure for a set number of clock cycles.

frequency
Definition of bus frequency details for f4.

i2c
Inter-Integrated Circuit (I2C)

led
User LED PA5

leds
Eight LEDs connected to PORTB

lsm9ds1
Interfacing the LSM9DS1 3D accelerometer, gyroscope, and magnetometer using SPI.

madgwick_ahrs
Rust implementation of Madgwick's IMU and AHRS algorithms.

math_utils
Some math utilities

prelude
The prelude is a collection of all the traits in this crate

pwm
Pulse Width Modulation

serial
Serial interface

spi
Serial Peripheral Interface

time
Units of time

timer
Timer

Macros

printf
printf over semihosting

println
println over semihosting

uprint
Macro for printing formatted strings over serial through DMA. Uses the cortex-m-rtfm resource model and can thus not be used outside rtfm tasks.