Struct f4::adc::Adc

pub struct Adc<'a>(pub &'a ADC1);

ADC1

Methods

impl<'a> Adc<'a>

pub fn enable_input(
&self,
input: AdcChannel,
sq: u8,
gpioa: &GPIOA,
gpiob: &GPIOB,
gpioc: &GPIOC
)

Enables the ADC input

pub fn init(&self, dma2: &DMA2, rcc: &RCC)

Initializes the ADC

NOTE Pwm<TIM2>.init must be called before this method because both methods configure the PA1 pin (one as input and the other as output :-/)

pub fn disable(&self)

Disables the ADC

pub fn enable(&self)

Enables the ADC

pub fn start<B>(
&self,
buffer: &Static<CircBuffer<B, Dma2Stream0>>,
dma2: &DMA2,
pwm: Pwm<TIM2>
) -> Result<(), Error> where
B: Unsize<[u16]>,

Starts an analog to digital conversion that will be periodically triggered by the channel 2 of TIM2

The conversions will be stored in the circular buffer