Module f4::dma

Direct Memory Access (DMA)

Structs

Buffer
Buffer to be used with a certain DMA STREAM

CircBuffer
A circular buffer associated to a DMA STREAM

Dma1Stream1
Stream 1 of DMA1

Dma1Stream2
Stream 2 of DMA1

Dma1Stream4
Stream 4 of DMA1

Dma1Stream5
Stream 5 of DMA1

Dma1Stream6
Stream 6 of DMA1

Dma2Stream0
Stream 0 of DMA2

Ref
Wraps a borrowed reference to a value in a Buffer

RefMut
A wrapper type for a mutably borrowed value from a `Buffer``

Enums

Error
DMA error