Struct f4::dma::CircBuffer

pub struct CircBuffer<B, STREAM> { /* fields omitted */ }

A circular buffer associated to a DMA STREAM

Methods

impl<B, STREAM> CircBuffer<B, STREAM>

pub const fn new(buffer: [B; 2]) -> Self

Constructs a circular buffer from two halves

impl<B> CircBuffer<B, Dma1Stream1>

pub fn read<R, F>(&self, dma1: &DMA1, f: F) -> Result<R, Error> where
F: FnOnce(&B) -> R,

Yields read access to the half of the circular buffer that's not currently being mutated by the DMA

impl<B> CircBuffer<B, Dma2Stream0>

pub fn read<R, F>(&self, dma2: &DMA2, f: F) -> Result<R, Error> where
F: FnOnce(&B) -> R,

Yields read access to the half of the circular buffer that's not currently being mutated by the DMA