Trait f4::prelude::_embedded_hal_Capture

pub trait _embedded_hal_Capture {
    type Error;
    type Channel;
    type Time;
    type Capture;
    fn capture(
        &self, 
        channel: Self::Channel
    ) -> Result<Self::Capture, Error<Self::Error>>;

    fn disable(&self, channel: Self::Channel);

    fn enable(&self, channel: Self::Channel);

    fn get_resolution(&self) -> Self::Time;

    fn set_resolution<R>(&self, resolution: R)
    where
        R: Into<Self::Time>;
}

Input capture

Examples

You can use this interface to measure the period of (quasi) periodic signals / events

let capture: impl Capture = ..;

capture.set_resolution(1.ms());

let before = block!(capture.capture()).unwrap();
let after = block!(capture.capture()).unwrap();

let period = after.wrapping_sub(before);

println!("Period: {} ms", period);

Associated Types

type Error

Enumeration of Capture errors

Possible errors:

overcapture, the previous capture value was overwritten because it was not read in a timely manner

type Channel

Enumeration of channels that can be used with this Capture interface

If your Capture interface has no channels you can use the type () here

type Time

A time unit that can be converted into a human time unit (e.g. seconds)

type Capture

The type of the value returned by capture

Required Methods

fn capture(
&self,
channel: Self::Channel
) -> Result<Self::Capture, Error<Self::Error>>

"Waits" for a transition in the capture channel and returns the value of counter at that instant

NOTE that you must multiply the returned value by the resolution of this Capture interface to get a human time unit (e.g. seconds)

fn disable(&self, channel: Self::Channel)

Disables a capture channel

fn enable(&self, channel: Self::Channel)

Enables a capture channel

fn get_resolution(&self) -> Self::Time

Returns the current resolution

fn set_resolution<R>(&self, resolution: R) where
R: Into<Self::Time>,

Sets the resolution of the capture timer