Struct f4::pwm::Pwm

pub struct Pwm<'a, T>(pub &'a T)
 where
    T: 'a;

PWM driver

Methods

impl<'a> Pwm<'a, TIM1>

pub fn init<P>(
&self,
period: P,
channel: Channel,
dma1: Option<&DMA1>,
gpioa: &GPIOA,
gpiob: &GPIOB,
gpioc: &GPIOC,
rcc: &RCC
) where
P: Into<Ticks>,

Initializes the PWM module

pub fn set_duties<B>(
&self,
dma1: &DMA1,
channel: Channel,
buffer: &Static<Buffer<B, Dma1Stream2>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Uses buffer to continuously change the duty cycle on every period

impl<'a> Pwm<'a, TIM2>

pub fn init<P>(
&self,
period: P,
channel: Channel,
dma1: Option<&DMA1>,
gpioa: &GPIOA,
gpiob: &GPIOB,
gpioc: &GPIOC,
rcc: &RCC
) where
P: Into<Ticks>,

Initializes the PWM module

pub fn set_duties<B>(
&self,
dma1: &DMA1,
channel: Channel,
buffer: &Static<Buffer<B, Dma1Stream2>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Uses buffer to continuously change the duty cycle on every period

impl<'a> Pwm<'a, TIM3>

pub fn init<P>(
&self,
period: P,
channel: Channel,
dma1: Option<&DMA1>,
gpioa: &GPIOA,
gpiob: &GPIOB,
gpioc: &GPIOC,
rcc: &RCC
) where
P: Into<Ticks>,

Initializes the PWM module

pub fn set_duties<B>(
&self,
dma1: &DMA1,
channel: Channel,
buffer: &Static<Buffer<B, Dma1Stream2>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Uses buffer to continuously change the duty cycle on every period

impl<'a> Pwm<'a, TIM4>

pub fn init<P>(
&self,
period: P,
channel: Channel,
dma1: Option<&DMA1>,
gpioa: &GPIOA,
gpiob: &GPIOB,
gpioc: &GPIOC,
rcc: &RCC
) where
P: Into<Ticks>,

Initializes the PWM module

pub fn set_duties<B>(
&self,
dma1: &DMA1,
channel: Channel,
buffer: &Static<Buffer<B, Dma1Stream2>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Uses buffer to continuously change the duty cycle on every period