Struct f4::serial::Serial

pub struct Serial<'a, U>(pub &'a U)
 where
    U: Any + Usart;

Serial interface

Interrupts

RXNE

Methods

impl<'a, U> Serial<'a, U> where
U: Any + Usart,

pub fn init<B>(
&self,
baud_rate: B,
dma1: Option<&DMA1>,
gpio: &U::GPIO,
rcc: &RCC
) where
B: Into<U::Ticks>,

Initializes the serial interface with a baud rate of baut_rate bits per second

The serial interface will be configured to use 8 bits of data, 1 stop bit, no hardware control and to omit parity checking

pub fn listen(&self, event: Event)

Starts listening for an interrupt event

pub fn unlisten(&self, event: Event)

Stops listening for an interrupt event

impl<'a> Serial<'a, USART2>

pub fn read_exact<B>(
&self,
dma1: &DMA1,
buffer: &Static<Buffer<B, Dma1Stream5>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Starts a DMA transfer to receive serial data into a buffer

This will mutably lock the buffer preventing borrowing its contents The buffer can be released after the DMA transfer finishes

pub fn write_all<B>(
&self,
dma1: &DMA1,
buffer: &Static<Buffer<B, Dma1Stream6>>
) -> Result<(), Error> where
B: Unsize<[u8]>,

Starts a DMA transfer to send buffer through this serial port

This will immutably lock the buffer preventing mutably borrowing its contents. The buffer can be released after the DMA transfer finishes