Struct f4::timer::Timer

pub struct Timer<'a, T>(pub &'a T)
 where
    T: 'a;

hal::Timer implementation

Methods

impl<'a> Timer<'a, TIM1>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM2>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM3>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM4>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM5>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM9>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM10>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.

impl<'a> Timer<'a, TIM11>

pub fn init<P>(&self, period: P, rcc: &RCC) where
P: Into<Ticks>,

Initializes the timer with a periodic timeout of frequency Hz

NOTE After initialization, the timer will be in the paused state.