Struct f4::dma::Buffer

#[repr(packed)]

pub struct Buffer<T, CHANNEL> { /* fields omitted */ }

Buffer to be used with a certain DMA CHANNEL

Methods

impl<T, CHANNEL> Buffer<T, CHANNEL>

pub const fn new(data: T) -> Self

Creates a new buffer

pub fn borrow(&self) -> Ref<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed.

impl<T> Buffer<T, Dma1Channel2>

pub fn release(&self, dma1: &DMA1) -> Result<(), Error>

Waits until the DMA releases this buffer

impl<T> Buffer<T, Dma1Channel4>

pub fn release(&self, dma1: &DMA1) -> Result<(), Error>

Waits until the DMA releases this buffer

impl<T> Buffer<T, Dma1Channel5>

pub fn release(&self, dma1: &DMA1) -> Result<(), Error>

Waits until the DMA releases this buffer

impl<T> Buffer<T, Dma1Channel6>

pub fn release(&self, dma1: &DMA1) -> Result<(), Error>

Waits until the DMA releases this buffer